separated out actions, implemented better model
