+++ src/actions/__init__.py
@@ -3,13 +3,37 @@
 # -*- coding: utf-8 -*-
 """
 pi_control
-__init__.py.py
+actions/__init__.py
 Author: Danyal Ahsanullah
 Date: 6/25/2018
 Copyright (c):  2018 Danyal Ahsanullah
 License: N/A
-Description: 
+Description:
 """
+# prototype action function signature:
+#
+# def some_action(interface: obj, params: dict, nxt: dict) -> Callable:
+#     """
+#     Action description
+#     ...
+#     :param interface: base interface abstraction layer that is performing an action.
+#     :param params: dictionary of the form
+#     	{'param0':<val>, 'param1':<val>, ..., 'paramN':<val>}
+#     :param nxt: dictionary of state transitions in the form
+#     	{<condition>:<action_func>, <condition>:<action_func>, None:<action_func>}
+#     """
+#     condition = None
+#     # stuff happens here -- eval and assign the desired value for condition
+#     ...
+#     return next[condition]
+#
 
-if __name__ == '__main__':
-    pass
+from .calibrate import calibrate
+from .reset_max import reset_max
+from .reset_min import reset_min
+
+actions = {'RESET_MIN': 'reset_min',
+           'RESET_MAX': 'reset_max',
+           'GOTO_POS': 'set_pos',
+           'CLEANUP': 'cleanup',
+           }
