+++ src/diagnostic_script.py
@@ -2,26 +2,13 @@
 
 # TODO: check if refactor works
 
-import os
 from collections import deque
-from time import (
-    strftime,
-    perf_counter,
-    # sleep,
-)
-# noinspection PyUnresolvedReferences
-from libs.utils import cfg_formatter, get_k_value
-from libs.controller import (
-    PController,
-    PDController,
-    PIController,
-    PIDController
-)
-from launch.cli_parser import (
-    parser,
-    cmds,
-    actions,
-)
+from time import strftime, perf_counter  # , sleep
+
+from actions import actions
+from launch import parser, cmds
+from libs.utils import cfg_formatter
+from actions.calibrate import calibrate
 from libs.hal import (
     adc,
     dac,
@@ -31,7 +18,6 @@ from libs.hal import (
     actuator,
     hal_init,
     hal_cleanup,
-
 )
 
 # globals for routines & ISRs
@@ -86,190 +72,6 @@ def monitor_adc_isr(channel):
         actuator.set_actuator_dir('forward')
         print('forward')
         # GPIO.output(PINS['relay_1'], 1)
-
-
-# routines
-####################################################
-# general use routines
-def reset_max():
-    actuator.set_position(actuator.pos_limit_high)
-
-
-def reset_min():
-    actuator.set_position(actuator.pos_limit_low)
-
-
-# calibration routines
-def calibrate_position():
-    """
-    hand calibrate system position thresholds
-    """
-    print('Beginning calibration routine...')
-    # TODO: figure out what relay 2 does
-    actuator.set_actuator_dir('forward')
-    dac.set_voltage(dac.stop)
-    print('Setting upper threshold')
-    input('Hit enter/return to begin.')
-    dac.set_voltage(dac.default_val)
-    input('Hit enter/return to mark absolute upper threshold')
-    actuator.pos_limit_high = adc.read_adc(adc.default_channel, gain=adc.gain, data_rate=adc.sample_rate)
-    dac.set_voltage(dac.stop)
-    print(actuator.pos_limit_high)
-    GPIO.output(PINS['relay_1'], 0)  # prepare to go backwards
-    print('Setting lower threshold')
-    input('Hit enter/return to begin.')
-    dac.set_voltage(dac.default_val)
-    input('Hit enter/return to mark absolute lower threshold')
-    actuator.pos_limit_low = adc.read_adc(adc.default_channel, gain=adc.gain, data_rate=adc.sample_rate)
-    dac.set_voltage(dac.stop)
-    print(actuator.pos_limit_low)
-    GPIO.output(PINS['relay_1'], 1)  # prepare to go forward
-    print('Setting upper desired threshold')
-    input('Hit enter/return  to begin.')
-    dac.set_voltage(dac.default_val)
-    input('Hit enter/return to mark desired upper threshold')
-    actuator.pos_threshold_high = adc.read_adc(adc.default_channel, gain=adc.gain, data_rate=adc.sample_rate)
-    dac.set_voltage(dac.stop)
-    print(actuator.pos_threshold_high)
-    GPIO.output(PINS['relay_1'], 0)  # prepare to go backwards
-    print('Setting lower desired threshold')
-    input('Hit enter/return  to begin.')
-    dac.set_voltage(dac.default_val)
-    input('Hit enter/return to mark desired lower threshold')
-    actuator.pos_threshold_low = adc.read_adc(adc.default_channel, gain=adc.gain, data_rate=adc.sample_rate)
-    dac.set_voltage(dac.stop)
-    print(actuator.pos_threshold_low)
-
-
-def set_controller():
-    control_map = {1: None, 2: PController, 3: PDController, 4: PIController, 5: PIDController}
-    global CONTROLLER
-    spaces = ' ' * 5
-    valid_choices = set(map(str, control_map.keys()))
-    print('set desired controller')
-    print('valid choices:')
-    print(spaces + '\n{}'.format(spaces).join(['{!s}: {!s}'.format(k, getattr(v, '__name__', v))
-                                               for k, v in control_map.items()]))
-    choice = input('Enter controller choice: ').strip()
-    while choice not in valid_choices:  # delay int conversion to handle invalid string inputs
-        choice = input('Enter controller choice: ').strip()
-    controller = control_map[int(choice)]
-    if controller is None:
-        return
-    coefficients = {k: 0.0 for k in controller.coefficients}
-    for coefficient in controller.coefficients:
-        coefficients[coefficient] = get_k_value(coefficient)
-    CONTROLLER = controller(**coefficients)
-
-
-def test_configurations():
-    flag = False
-    print('Testing current config of:')
-    print('Controller: {!s}'.format(CONTROLLER))
-    print('Position Thresholds:')
-    print('Absolute low: {}'.format(actuator.pos_limit_low))
-    print('Absolute high: {}'.format(actuator.pos_limit_high))
-    print('Set low: {}'.format(actuator.pos_threshold_low))
-    print('Set High: {}'.format(actuator.pos_threshold_high))
-    actuator.set_actuator_dir('forward')
-    # GPIO.output(PINS['relay_1'], 1)  # set dir as forward
-    dac.value = 1024
-    dac.set_voltage(dac.value)
-    value = adc.start_adc_comparator(adc.default_channel, 2 ** 16 - 1, 0, gain=adc.gain, data_rate=adc.sample_rate)
-    while not flag:
-        if value >= actuator.pos_limit_high:
-            print('Hit absolute max limit', value)
-            actuator.set_actuator_dir('backward')
-            # GPIO.output(PINS['relay_1'], 0)
-            flag = True
-        # GPIO.wait_for_edge(PINS['adc_alert'], GPIO.FALLING, timeout=WAIT_TIMEOUT)
-        adc.wait_for_sample(timeout=WAIT_TIMEOUT)
-        value = adc.get_last_result()
-    flag = False
-    value = adc.get_last_result()
-    while not flag:
-        if value <= actuator.pos_limit_low:
-            print('Hit absolute min limit', value)
-            actuator.set_actuator_dir('forward')
-            # GPIO.output(PINS['relay_1'], 1)
-            flag = True
-        # GPIO.wait_for_edge(PINS['adc_alert'], GPIO.FALLING, timeout=WAIT_TIMEOUT)
-        adc.wait_for_sample(timeout=WAIT_TIMEOUT)
-        value = adc.get_last_result()
-    flag = False
-    value = adc.get_last_result()
-    while not flag:
-        if value >= actuator.pos_threshold_high:
-            print('Hit designated max limit', value)
-            actuator.set_actuator_dir('backward')
-            # GPIO.output(PINS['relay_1'], 0)
-            flag = True
-            # adc.stop_adc()
-        # GPIO.wait_for_edge(PINS['adc_alert'], GPIO.FALLING, timeout=WAIT_TIMEOUT)
-        adc.wait_for_sample(timeout=WAIT_TIMEOUT)
-        value = adc.get_last_result()
-    flag = False
-    value = adc.get_last_result()
-    while not flag:
-        if value <= actuator.pos_threshold_low:
-            dac.set_voltage(dac.stop)
-            print('Hit designated min limit', value)
-            actuator.set_actuator_dir('forward')
-            # GPIO.output(PINS['relay_1'], 1)
-            flag = True
-        # GPIO.wait_for_edge(PINS['adc_alert'], GPIO.FALLING, timeout=WAIT_TIMEOUT)
-        adc.wait_for_sample(timeout=WAIT_TIMEOUT)
-        value = adc.get_last_result()
-    dac.set_voltage(dac.stop)
-    adc.stop_adc()
-
-
-def calibrate():
-    """
-    calibrate actuator control
-    allows for :
-        thresholding:
-            software implemented stroke boundaries
-            safety and desired operating stroke definitions
-        control method:
-            P control - output speed is proportional to desired position to approach that position
-            PD control - P control + output speed is modified by the rate of approach
-            None - output speed is flat over entire duration
-        testing and editing of generated parameters:
-            test currently generated scheme
-            editing any single point generated
-
-    7 step process:
-        1. define abs low (leave alone if limit switch)
-        2. define abs high (leave alone if limit switch)
-        3. define desired stroke low threshold
-        4. define desired stroke high threshold
-        5. define controller
-        6. define desired actuator movement rates
-        6.5. test current setup -- tweak values as wanted
-        7. confirm cal data
-    """
-    # calibrate thresholds:
-    calibrate_position()
-    # set control scheme:
-    set_controller()
-    # test:
-    test_configurations()
-    # reconfigure -- optional
-    # edit = input('edit configuration? (y/n): ').strip().lower()
-    # while edit not in ('y','n'):
-    #     edit = input('edit configuration? (y/n): ').strip().lower()
-    # if edit == 'y':
-    #     edit_config()
-    # confirm
-    confirm = input('confirm settings? (y/n): ').strip().lower()
-    while confirm not in ('y', 'n'):
-        confirm = input('confirm settings? (y/n): ').strip().lower()
-    if confirm == 'y':
-        outfile = input('enter valid config file name: ').strip()
-        while not os.access(os.path.dirname(outfile), os.W_OK):
-            outfile = input('enter valid config file name: ').strip()
-        save_config(outfile)
 
 
 # diagnostic_routines
@@ -378,16 +180,16 @@ def dispatcher(arg_dict):
     elif arg_dict['cmd'] == cmds['TEST_DAC']:
         test_dac(**arg_dict)
     elif arg_dict['cmd'] == cmds['TEST_CAL']:
-        calibrate()
+        calibrate(interface=actuator, params={}, nxt=None)
     elif arg_dict['cmd'] == cmds['TEST_POS']:
         actuator.pos_limit_low = arg_dict.pop('low_min')
         actuator.pos_limit_high = arg_dict.pop('high_max')
         actuator.pos_threshold_low = arg_dict.pop('low_threshold')
         actuator.pos_threshold_high = arg_dict['high_threshold']
         if arg_dict['action'] == actions['RESET_MIN']:
-            reset_min()
+            actuator.reset_min()
         elif arg_dict['action'] == actions['RESET_MAX']:
-            reset_max()
+            actuator.reset_max()
         elif arg_dict['action'] == actions['GOTO_POS']:
             actuator.set_position(arg_dict['position'])
     elif arg_dict['cmd'] == cmds['RUN_ACQ']:
