+++ src/launch/__init__.py
@@ -3,10 +3,12 @@
 # -*- coding: utf-8 -*-
 """
 pi_control
-__init__.py.py
+launch/__init__.py
 Author: Danyal Ahsanullah
 Date: 6/15/2018
 Copyright (c):  2018 Danyal Ahsanullah
 License: N/A
 Description: 
 """
+from .cli_parser import parser
+from .command_map import cmds
