+++ src/launch/cli_parser.py
@@ -20,27 +20,12 @@ Usage:
         ...
 """
 
-import sys
-import yaml
 from argparse import ArgumentParser
 
+from actions import actions
+from launch.command_map import cmds
 from version import version as __version__
-from libs.hal import adc, set_config, actuator
-from routines import register_routine, register_action
-
-# command names:
-cmds = {'TEST_ADC': 'test_adc',
-        'TEST_DAC': 'test_dac',
-        'TEST_CAL': 'test_cal',
-        'TEST_POS': 'test_pos',
-        'RUN_ACQ': 'acquire',
-        }
-
-actions = {'RESET_MIN': 'reset_min',
-           'RESET_MAX': 'reset_max',
-           'GOTO_POS': 'set_pos',
-           'CLEANUP': 'cleanup',
-           }
+from libs.hal import adc, actuator
 
 parser = ArgumentParser()  # 'elastocaloric testing'
 subparsers = parser.add_subparsers(help='Action to take', dest='cmd')
@@ -86,77 +71,6 @@ monitor_parser.add_argument('-h', '--high_threshold', type=int, default=actuator
 monitor_parser.add_argument('-H', '--high_max', type=int, default=actuator.pos_limit_high)
 monitor_parser.add_argument('--help', action='help', help='print help')
 
-DOC_DISPATCHER = {
-    'CFG': set_config,
-    'RTN': register_routine,
-    'ACT': register_action,
-}
-
-
-def load_config(path):
-    with open(path, 'r') as file:
-        docs = yaml.load_all(file)
-        cfg = {}
-        for doc in docs:
-            cfg = eval(repr(doc))
-            # DOC_DISPATCHER[doc.__type](doc)
-        return cfg
-
-
-class ReprMixIn:
-    def __repr__(self):
-        return '{!s}({!s})'.format(self.__class__.__name__,
-                                   ', '.join('{!s}={!r}'.format(k, v) for k, v in vars(self).items()))
-
-
-# TODO: figure out validation -- looks like have to define a method in the from_yaml hook
-class Config(yaml.YAMLObject, ReprMixIn):
-    yaml_tag = '!Config'
-    __type = 'CFG'
-    accepted_units = {'raw', 'imperial', 'metric'}  # raw: A/D levels, imperial: in, lb, metric: mm, kg
-    accepted_adc_sample_rates = adc.accepted_sample_rates
-    accepted_adc_gains = adc.accepted_gains
-    accepted_adc_channels = adc.accepted_channels
-
-    def __init__(self, version, units, upper_limit, lower_limit, adc_sample_rate, adc_gain, adc_channel):
-        if units in self.accepted_units and \
-                adc_sample_rate in self.accepted_adc_sample_rates and \
-                adc_gain in self.accepted_adc_gains and \
-                adc_channel in self.accepted_adc_channels:
-            self.version = version  # to be used for future releases
-            self.units = units  # units to use for input/output values. internally math is done independent of units
-            self.upper_limit = upper_limit  # upper limit for actuator, stored and used in calculations as raw adc level
-            self.lower_limit = lower_limit  # lower limit for actuator, stored and used in calculations as raw adc level
-            self.adc_sample_rate = adc_sample_rate  # configurable, value must be: 8, 16, 32, 64, 128, 250, 475, 860
-            self.adc_gain = adc_gain  # configurable, value must be: 2/3, 1, 2, 4, 8, 16
-            self.adc_channel = adc_channel  # configurable, value must be: 0, 1, 2, 3
-        else:
-            raise ValueError('Bad key in yaml configuration.')
-
-
-class Routine(yaml.YAMLObject, ReprMixIn):
-    yaml_tag = '!Routine'
-    __type = 'RTN'
-
-    def __init__(self, name, units, actions, output=sys.stdout):
-        self.name = name
-        self.units = units
-        self.actions = actions
-        self.output = output
-
-
-class Action(yaml.YAMLObject, ReprMixIn):
-    yaml_tag = '!Action'
-    __type = 'ACT'
-
-    def __init__(self, name, params=None):
-        self.action = actions[name]
-        self.params = params
-
 
 if __name__ == '__main__':
-    from pprint import pprint
-
     parser.print_help()
-    print('')
-    pprint(load_config('test_config.yaml'))
