+++ src/launch/command_map.py
@@ -1,0 +1,21 @@
+#! /usr/bin/env python
+# vim:fileencoding=utf-8
+# -*- coding: utf-8 -*-
+"""
+pi_control
+command_map.py
+Author: Danyal Ahsanullah
+Date: 6/26/2018
+Copyright (c):  2018 Danyal Ahsanullah
+License: N/A
+Description: 
+"""
+
+if __name__ == '__main__':
+    pass
+cmds = {'TEST_ADC': 'test_adc',
+        'TEST_DAC': 'test_dac',
+        'TEST_CAL': 'test_cal',
+        'TEST_POS': 'test_pos',
+        'RUN_ACQ': 'acquire',
+        }
