+++ src/routines/__init__.py
@@ -3,18 +3,10 @@
 # -*- coding: utf-8 -*-
 """
 pi_control
-__init__.py.py
+routines/__init__.py
 Author: Danyal Ahsanullah
 Date: 6/11/2018
 Copyright (c):  2018 Danyal Ahsanullah
 License: N/A
 Description: import available routines here
 """
-
-
-def register_routine():
-    return None
-
-
-def register_action():
-    return None
