+++ test/test_config.yaml
@@ -1,0 +1,43 @@
+# global configuration parameters
+CONFIG: &CONFIG !Config
+  version: 0.0.1
+  units: raw &UNITS
+  lower_limit: &LIMIT_LOW 0
+  upper_limit: &LIMIT_HIGH 20000
+  adc_sample_rate: 860
+  adc_gain: 1
+  adc_channel: 1
+# define any actions that you want to reuse multiple times in multiple routines
+# ensure you overwrite the nxt table on each usage if needed
+ACTIONS: &ACTIONS
+  - !Action &RESET_MAX
+    name: RESET_MAX
+    params:
+      pos:
+        - 1
+        - 2
+        - 3
+    nxt:
+      True: RESET_MIN
+      failed: TEST
+      False: RESET_MAX  # loop back state
+# begin routine definitions
+ROUTINES: &ROUTINES
+  - !Routine
+    name: reset_min
+    units: &UNITS
+    actions:
+      - !Action
+          name: RESET_MIN
+          params:
+            low: *LIMIT_LOW
+            high: *LIMIT_HIGH
+          nxt:
+            True: CLEANUP
+            False: RESET_MIN  # loop back state -- completely useless usually
+  - !Routine
+    name: reset_max
+    units: raw
+    output: stdout
+    actions:
+      - *RESET_MAX
